+++ NEWS.md
@@ -1,5 +1,6 @@
 # explor (development version)
 
+- Fix duplicated level names in supplementary variables in speMCA
 - Fix warnings in Font Awesome icon names (#39, thanks @jl5000)
 
 # explor 0.3.9